correct type of a variable

+++ lib/widgets/select_log_value_sections.dart
@@ -296,7 +296,7 @@ class _ChooseDecimalLogValueSectionState
 
         try {
           // used to test whether the input is parable as a number
-          int parsedLogValue = int.parse(enteredValue);
+          double parsedLogValue = double.parse(enteredValue);
 
           if (_showInputNotParsableWarning) {
             // do not show the warning since the entered value is parsable
